+++ experimental/instrumentation-replay/jest.config.js
@@ -2,6 +2,6 @@ const { jestBaseConfig } = require('../../jest.config.base.js');
 
 module.exports = {
   ...jestBaseConfig,
-  roots: ['experimental/instrumentation-session-recording/src'],
+  roots: ['experimental/instrumentation-replay/src'],
   testEnvironment: 'jsdom',
 };