+++ experimental/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Next
 
-- Added support for `SessionRecordingInstrumentation`.
+- Added support for experimental `ReplayInstrumentation` (`@grafana/faro-instrumentation-replay`).
 
 ## 1.13.0